+++ app/src/main/java/com/vk/directop/memoriesjournal/core/domain/playback/AudioPlayer.kt
@@ -3,7 +3,7 @@ package com.vk.directop.memoriesjournal.core.domain.playback
 import java.io.File
 
 interface AudioPlayer {
-    fun playFile(file: File, onCompletion: () -> Unit, onProgress: (Int, Int) -> Unit)
+    fun playFile(file: File, onCompletion: () -> Unit, onProgress: (Long, Long) -> Unit)
     fun stop()
     fun getCurrentPosition(): Int
     fun getDuration(): Int
